XUnitTest begonnen

+++ UdemyNLayerProject.API/Controllers/ProductsController.cs
@@ -9,6 +9,7 @@ using System.Linq;
 using System.Threading.Tasks;
 
 using UdemyNLayerProject.API.DTOs;
+using UdemyNLayerProject.API.Exctentions;
 using UdemyNLayerProject.API.Filters;
 using UdemyNLayerProject.Core.Models;
 using UdemyNLayerProject.Core.Services;
@@ -21,6 +22,10 @@ namespace UdemyNLayerProject.API.Controllers
     {
         private readonly IMapper _mapper;
         private readonly IProductService _productService;
+        public ProductsController(IProductService productService)
+        {
+            _productService = productService;
+        }
         public ProductsController(IProductService productService, IMapper mapper)
         {
             _productService = productService;
@@ -31,6 +36,7 @@ namespace UdemyNLayerProject.API.Controllers
         public async Task<IActionResult> GetAll()
         {
             var products = await _productService.GetAllAsync();
+            if(products == null) { return NotFound(); }
             return Ok(_mapper.Map<IEnumerable<ProductDto>>(products));
         }
 
@@ -39,6 +45,7 @@ namespace UdemyNLayerProject.API.Controllers
         public async Task<IActionResult> GetById(int id)
         {
             var product = await _productService.GetByIdAsync(id);
+            if (product == null) { return NotFound(); }
             // 200
             return Ok(_mapper.Map<ProductDto>(product));
         }
@@ -47,14 +54,18 @@ namespace UdemyNLayerProject.API.Controllers
         [HttpGet("{id}/category")]
         public async Task<IActionResult> GetWithCategoryById(int id)
         {
-            var Product = await _productService.GetWithCategoryById(id);
+            var product = await _productService.GetWithCategoryById(id);
+            if (product == null) { return NotFound(); }
             // 200
-            return Ok(_mapper.Map<ProductWithCategory>(Product));
+            return Ok(_mapper.Map<ProductWithCategory>(product));
         }
 
         [HttpPost]
         public async Task<IActionResult> Save(ProductDto productDto)
         {
+            var v = await _productService.GetAllAsync();
+            var name = v.Select(x=>x.Name == productDto.Name).ToString();
+            if (name == productDto.Name) { return Conflict(); }//409
             var newProduct = await _productService.AddAsync(_mapper.Map<Product>(productDto));
             // 201
             return Created(string.Empty, _mapper.Map<ProductDto>(newProduct));
@@ -63,9 +74,14 @@ namespace UdemyNLayerProject.API.Controllers
         [HttpPut]
         public IActionResult Update(ProductDto productDto)
         {
+            var v =  _productService.GetAllAsync().Result;
+            //string id = v.Where(x=>x.Id == productDto.Id).FirstOrDefault().ToString();
+            if (productDto == null) { return BadRequest(); }
+            else if (v.Where(x => x.Id == productDto.Id).FirstOrDefault().ToString() == null) { return BadRequest(); }
             var product = _productService.Update(_mapper.Map<Product>(productDto));
             // 204
             return NoContent();
+
         }
 
         [ServiceFilter(typeof(NotFoundFilter))]
@@ -73,6 +89,7 @@ namespace UdemyNLayerProject.API.Controllers
         public IActionResult Remove(int id)
         {
             var product = _productService.GetByIdAsync(id).Result;
+            if (product == null) { return NotFound(); }
             _productService.Remove(product);
             // 204
             return NoContent();
